Refactor all sentry-related code to own file
